fix: プロジェクトの構造を変更
