+++ web/liff/src/layouts/default.vue
@@ -1,32 +1,74 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia';
+import liff from '@line/liff';
+import { useAuthStore } from '~/stores/auth';
 import { useShoppingCartStore } from '~/stores/shopping';
 
 const route = useRoute();
+const router = useRouter();
+const runtimeConfig = useRuntimeConfig();
 
-const facilityId = computed<string>(() => {
-  return String(route.query.facilityId || '');
-});
+const facilityId = computed<string>(() => String(route.query.facilityId || ''));
 
 const isExpand = ref<boolean>(false);
 
-// ショッピングカートストアを使用
+// ストア
 const shoppingCartStore = useShoppingCartStore();
+const authStore = useAuthStore();
 const { shoppingCart, cartIsEmpty, totalPrice, totalQuantity } = storeToRefs(shoppingCartStore);
 
 const toggleExpand = () => {
   isExpand.value = !isExpand.value;
 };
 
-// コンポーネントマウント時にカート情報を取得
+// マウント時に認証処理とカート取得を実行
 onMounted(async () => {
+  const liffId = runtimeConfig.public.LIFF_ID;
+  if (!liffId) {
+    console.error('Please set LIFF_ID in .env file');
+  }
+  else {
+    try {
+      console.log(liffId);
+      await liff.init({ liffId });
+    }
+    catch (error) {
+      console.error('LIFF init failed', error);
+    }
+  }
+
+  if (!liff.isLoggedIn()) {
+    liff.login();
+    return;
+  }
+
+  // サインイン（LIFFログイン済みの場合のみ）
+  if (liff.isLoggedIn()) {
+    const liffAccessToken = liff.getAccessToken();
+    if (liffAccessToken) {
+      try {
+        console.log('liffAccessToken:', liffAccessToken);
+        const res = await authStore.signIn(liffAccessToken);
+        if (!res?.userId) {
+          const current = router.resolve({ path: route.path, query: route.query }).href;
+          const query: Record<string, string> = { redirect: current };
+          if (facilityId.value) query.facilityId = facilityId.value;
+          await router.push({ path: '/checkin/new', query });
+          return;
+        }
+      }
+      catch (err) {
+        console.error('Auth signIn or redirect failed:', err);
+      }
+    }
+  }
+
+  // 認証処理の後にカート情報を取得（トークン付与のため）
   await shoppingCartStore.getCart();
 });
 
 // 価格のフォーマット
-const formatPrice = (price: number) => {
-  return price.toLocaleString('ja-JP');
-};
+const formatPrice = (price: number) => price.toLocaleString('ja-JP');
 </script>
 
 <template>
