+++ web/liff/src/pages/[facilityId]/checkin/index.vue
