+++ web/liff/src/pages/[facilityId]/checkin/new.vue
