+++ web/liff/src/pages/[facilityId]/items/[id].vue
